Complete ACL.
 On branch upgrade_5
 Your branch is up to date with 'origin/upgrade_5'.

 Changes to be committed:
	modified:   composer.json
	modified:   composer.lock
	modified:   config/app.php
	new file:   config/auth_acl.ini
	new file:   config/auth_allow.ini
	modified:   config/bootstrap.php
	modified:   config/plugins.php
	modified:   config/routes.php
	modified:   plugins/Frontend/templates/Posts/view.php
	modified:   plugins/Frontend/templates/SystemicPages/xml/sitemap.php
	modified:   plugins/Frontend/templates/Users/login.php
	modified:   plugins/Frontend/templates/cell/PostCategories/display.php
	modified:   plugins/Frontend/templates/cell/Posts/last.php
	modified:   plugins/Frontend/templates/cell/Posts/popular.php
	modified:   plugins/Frontend/templates/cell/Posts/same_post_category.php
	modified:   plugins/Frontend/templates/element/Posts/post.php
	modified:   plugins/Frontend/templates/element/headers/nav.php
	modified:   plugins/Panel/config/app_form.php
	modified:   plugins/Panel/src/View/Helper/EmailAddressesHelper.php
	modified:   plugins/Panel/src/View/Helper/FilesHelper.php
	modified:   plugins/Panel/src/View/Helper/PanelHelper.php
	modified:   plugins/Panel/src/View/Helper/PhoneNumbersHelper.php
	modified:   plugins/Panel/templates/Admin/Faqs/add.php
	modified:   plugins/Panel/templates/Admin/Faqs/edit.php
	modified:   plugins/Panel/templates/Admin/Faqs/index.php
	modified:   plugins/Panel/templates/Admin/Faqs/translate.php
	modified:   plugins/Panel/templates/Admin/Files/add.php
	modified:   plugins/Panel/templates/Admin/Files/add_open_graph.php
	modified:   plugins/Panel/templates/Admin/Files/documents.php
	modified:   plugins/Panel/templates/Admin/Files/images.php
	modified:   plugins/Panel/templates/Admin/Files/index.php
	modified:   plugins/Panel/templates/Admin/Files/open_graph.php
	deleted:    plugins/Panel/templates/Admin/Files/videos.php
	modified:   plugins/Panel/templates/Admin/Pages/add.php
	modified:   plugins/Panel/templates/Admin/Pages/edit.php
	modified:   plugins/Panel/templates/Admin/Pages/index.php
	modified:   plugins/Panel/templates/Admin/Pages/translate.php
	modified:   plugins/Panel/templates/Admin/PostCategories/add.php
	modified:   plugins/Panel/templates/Admin/PostCategories/edit.php
	modified:   plugins/Panel/templates/Admin/PostCategories/index.php
	modified:   plugins/Panel/templates/Admin/PostCategories/translate.php
	modified:   plugins/Panel/templates/Admin/Posts/add.php
	modified:   plugins/Panel/templates/Admin/Posts/edit.php
	modified:   plugins/Panel/templates/Admin/Posts/index.php
	modified:   plugins/Panel/templates/Admin/Posts/translate.php
	modified:   plugins/Panel/templates/Admin/Services/add.php
	modified:   plugins/Panel/templates/Admin/Services/edit.php
	modified:   plugins/Panel/templates/Admin/Services/index.php
	modified:   plugins/Panel/templates/Admin/Services/translate.php
	modified:   plugins/Panel/templates/Admin/Settings/add.php
	modified:   plugins/Panel/templates/Admin/Settings/by_key.php
	modified:   plugins/Panel/templates/Admin/Settings/index.php
	modified:   plugins/Panel/templates/Admin/SystemicPages/add.php
	modified:   plugins/Panel/templates/Admin/SystemicPages/dashboard.php
	modified:   plugins/Panel/templates/Admin/SystemicPages/edit.php
	modified:   plugins/Panel/templates/Admin/SystemicPages/index.php
	modified:   plugins/Panel/templates/Admin/SystemicPages/translate.php
	modified:   plugins/Panel/templates/Admin/element/MetaTags/default.php
	modified:   plugins/Panel/templates/Admin/element/MetaTags/original.php
	modified:   plugins/Panel/templates/Admin/element/MetaTags/translate.php
	modified:   plugins/Panel/templates/Admin/element/breadcrumbs.php
	modified:   plugins/Panel/templates/Admin/element/change_password.php
	modified:   plugins/Panel/templates/Admin/element/flash/error.php
	modified:   plugins/Panel/templates/Admin/element/flash/success.php
	modified:   plugins/Panel/templates/Admin/element/footer.php
	modified:   plugins/Panel/templates/Admin/element/navigation.php
	modified:   plugins/Panel/templates/layout/default.php
	new file:   resources/locales/admin.pot
	new file:   resources/locales/default.pot
	new file:   resources/locales/frontend.pot
	new file:   resources/locales/ru/frontend.mo
	new file:   resources/locales/ru/frontend.po
	modified:   src/Application.php
	modified:   src/Controller/Admin/FilesController.php
	modified:   src/Controller/Admin/I18nMessagesController.php
	modified:   src/Controller/Admin/SettingsController.php
	modified:   src/Controller/AppController.php
	modified:   src/Controller/Component/ChangePasswordComponent.php
	modified:   src/Controller/PagesController.php
	modified:   src/Controller/PostsController.php
	modified:   src/Controller/SystemicPagesController.php
	modified:   src/Controller/UsersController.php
	new file:   src/Model/Entity/Role.php
	new file:   src/Model/Entity/RolesUser.php
	modified:   src/Model/Entity/User.php
	new file:   src/Model/Table/RolesTable.php
	new file:   src/Model/Table/RolesUsersTable.php
	modified:   src/Model/Table/UsersTable.php
	modified:   src/View/AppView.php
	new file:   tests/Fixture/RolesFixture.php
	new file:   tests/Fixture/RolesUsersFixture.php
	new file:   tests/TestCase/Model/Table/RolesTableTest.php
	new file:   tests/TestCase/Model/Table/RolesUsersTableTest.php

+++ plugins/Panel/templates/Admin/Faqs/edit.php
@@ -55,15 +55,17 @@ $(document).ready(function() {
                 <h2><?= __d('admin', 'Edit FAQ') ?></h2>
                 <div class="panel-toolbar ml-auto mr-3">
                     <?php
-                    echo $this->Form->deleteLink(
-                        $this->Html->tag('i', '', ['class' => 'fal fa-trash']) . ' ' . __d('panel', 'Delete'),
-                        ['controller' => 'Faqs', 'action' => 'delete', h($faq->id)],
-                        [
-                            'class' => 'btn btn-danger btn-xs mr-auto',
-                            'confirm' => __d('panel', 'Are you sure you want to delete this FAQ?'),
-                            'escape' => false
-                        ]
-                    );
+                    if ($this->AuthUser->hasAccess(['controller' => 'Faqs', 'action' => 'delete', h($faq->id)])) {
+                        echo $this->Form->deleteLink(
+                            $this->Html->tag('i', '', ['class' => 'fal fa-trash']) . ' ' . __d('admin', 'Delete'),
+                            ['controller' => 'Faqs', 'action' => 'delete', h($faq->id)],
+                            [
+                                'class' => 'btn btn-danger btn-xs mr-auto',
+                                'confirm' => __d('admin', 'Are you sure you want to delete this FAQ?'),
+                                'escape' => false
+                            ]
+                        );
+                    }
                     ?>
                 </div>
             </div>
@@ -76,15 +78,15 @@ $(document).ready(function() {
                             echo $this->Form->control('question', [
                                 'label' => __d('admin', 'Question') . $this->Html->tag('span' , '*', ['class' => 'ml-1 text-danger']),
                                 'escape' => false,
-                                'placeholder' => __d('panel', 'Question')
+                                'placeholder' => __d('admin', 'Question')
                             ]);
                             echo $this->Form->control('slug', [
                                 'label' => __d('admin', 'Slug') . $this->Html->tag('span' , '*', ['class' => 'ml-1 text-danger']),
                                 'escape' => false,
-                                'placeholder' => __d('panel', 'Slug')
+                                'placeholder' => __d('admin', 'Slug')
                             ]);
                             echo $this->Form->control('service_id', [
-                                'empty' => __d('panel', 'Select the service'),
+                                'empty' => __d('admin', 'Select the service'),
                                 'label' => __d('admin', 'Service') . $this->Html->tag('span' , '*', ['class' => 'ml-1 text-danger']),
                                 'escape' => false,
                                 'class' => 'form-control select2 w-100'
@@ -93,7 +95,7 @@ $(document).ready(function() {
                                 'label' => __d('admin', 'Answer') . $this->Html->tag('span' , '*', ['class' => 'ml-1 text-danger']),
                                 'escape' => false,
                                 'rows' => 11,
-                                'placeholder' => __d('panel', 'Answer')
+                                'placeholder' => __d('admin', 'Answer')
                             ]);
                             ?>
                         </div>
@@ -105,12 +107,12 @@ $(document).ready(function() {
                         <div class="col-12">
                             <div class="border-top pt-3 text-right">
                                 <?php
-                                echo $this->Html->link(
+                                echo $this->AuthUser->link(
                                     __d('admin', 'Cancel'),
                                     ['controller' => 'Faqs', 'action' => 'index'],
                                     ['class' => 'btn btn-default mr-2']
                                 );
-                                echo $this->Form->submit(__d('panel', 'Save'));
+                                echo $this->Form->submit(__d('admin', 'Save'));
                                 ?>
                             </div>
                         </div>
