+++ plugins/Panel/templates/Admin/Pages/edit.php
@@ -4,7 +4,7 @@ $this->assign('title', h($page->title));
 $this->start('breadcrumbs');
 $breadcrumbs = [
     ['title' => __d('admin', 'Pages')],
-    ['title' => __d('panel', 'Dynamic Pages'), 'url' => ['controller' => 'Pages', 'action' => 'index']],
+    ['title' => __d('admin', 'Dynamic Pages'), 'url' => ['controller' => 'Pages', 'action' => 'index']],
     ['title' => h($page->title)]
 ];
 echo $this->element('breadcrumbs', ['breadcrumbs' => $breadcrumbs]);
@@ -73,15 +73,17 @@ $(document).ready(function() {
                 <h2><?= __d('admin', 'Edit Dynamic Page') ?></h2>
                 <div class="panel-toolbar ml-auto mr-3">
                     <?php
-                    echo $this->Form->deleteLink(
-                        $this->Html->tag('i', '', ['class' => 'fal fa-trash']) . ' ' . __d('panel', 'Delete'),
-                        ['controller' => 'Pages', 'action' => 'delete', h($page->id)],
-                        [
-                            'class' => 'btn btn-danger btn-xs mr-auto',
-                            'confirm' => __d('panel', 'Are you sure you want to delete this page?'),
-                            'escape' => false
-                        ]
-                    );
+                    if ($this->AuthUser->hasAccess(['controller' => 'Pages', 'action' => 'delete', h($page->id)])) {
+                        echo $this->Form->deleteLink(
+                            $this->Html->tag('i', '', ['class' => 'fal fa-trash']) . ' ' . __d('admin', 'Delete'),
+                            ['controller' => 'Pages', 'action' => 'delete', h($page->id)],
+                            [
+                                'class' => 'btn btn-danger btn-xs mr-auto',
+                                'confirm' => __d('admin', 'Are you sure you want to delete this page?'),
+                                'escape' => false
+                            ]
+                        );
+                    }
                     ?>
                 </div>
             </div>
@@ -94,12 +96,12 @@ $(document).ready(function() {
                             echo $this->Form->control('title', [
                                 'label' => __d('admin', 'Title') . $this->Html->tag('span' , '*', ['class' => 'ml-1 text-danger']),
                                 'escape' => false,
-                                'placeholder' => __d('panel', 'Title')
+                                'placeholder' => __d('admin', 'Title')
                             ]);
                             echo $this->Form->control('slug', [
                                 'label' => __d('admin', 'Slug') . $this->Html->tag('span' , '*', ['class' => 'ml-1 text-danger']),
                                 'escape' => false,
-                                'placeholder' => __d('panel', 'Slug')
+                                'placeholder' => __d('admin', 'Slug')
                             ]);
                             ?>
 
@@ -107,7 +109,7 @@ $(document).ready(function() {
                                 <div class="col-lg-6">
                                     <?php
                                     echo $this->Form->control('parent_id', [
-                                        'empty' => __d('panel', 'Select the parent'),
+                                        'empty' => __d('admin', 'Select the parent'),
                                         'label' => __d('admin', 'Parent'),
                                         'class' => 'form-control select2 w-100'
                                     ]);
@@ -116,7 +118,7 @@ $(document).ready(function() {
                                 <div class="col-lg-6">
                                     <?php
                                     echo $this->Form->control('header', [
-                                        'empty' => __d('panel', 'Select the header template'),
+                                        'empty' => __d('admin', 'Select the header template'),
                                         'label' => __d('admin', 'Header Template'),
                                         'options' => $this->Template->headerList()
                                     ]);
@@ -129,7 +131,7 @@ $(document).ready(function() {
                                 'label' => __d('admin', 'Body') . $this->Html->tag('span' , '*', ['class' => 'ml-1 text-danger']),
                                 'escape' => false,
                                 'rows' => 11,
-                                'placeholder' => __d('panel', 'Body')
+                                'placeholder' => __d('admin', 'Body')
                             ]);
                             ?>
                         </div>
@@ -141,12 +143,12 @@ $(document).ready(function() {
                         <div class="col-12">
                             <div class="border-top pt-3 text-right">
                                 <?php
-                                echo $this->Html->link(
+                                echo $this->AuthUser->link(
                                     __d('admin', 'Cancel'),
                                     ['controller' => 'Pages', 'action' => 'index'],
                                     ['class' => 'btn btn-default mr-2']
                                 );
-                                echo $this->Form->submit(__d('panel', 'Save'));
+                                echo $this->Form->submit(__d('admin', 'Save'));
                                 ?>
                             </div>
                         </div>
